+++ spark.json
@@ -1,5 +1,5 @@
 {
-	"name": "CarbonLibrary-WifiFork",
+	"name": "CarbonLibrary",
 	"author": "Garrett Bartley <wgb@wgb.me>",
 	"version": "0.0.5",
 	"license": "MIT"
